Attachments: move buttons
