+++ src/common/attachment-drafts/attachment-sources/ButtonAttachCamera.tsx
